requirements.txt
aiohttp==3.6.2
altgraph==0.17
asgiref==3.2.7
astroid==2.4.1
async-timeout==3.0.1
attrs==19.3.0
autopep8==1.5.2
backcall==0.2.0
bottle==0.12.18
bottle-websocket==0.2.9
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cryptography==2.9.2
cx-Freeze==6.2
decorator==4.4.2
discord.py==1.4.1
future==0.18.2
gevent==20.6.2
gevent-websocket==0.10.1
greenlet==0.4.16
idna==2.9
ipython==7.15.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.6
parso==0.7.0
pefile==2019.4.18
pickleshare==0.7.5
prompt-toolkit==3.0.5
py==1.8.1
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.6.1
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.7
pylint==2.5.2
pylotus==1.0.0
PySide2==5.15.0
pytz==2020.1
pywin32-ctypes==0.2.0
requests==2.23.0
shiboken2==5.15.0
six==1.14.0
sqlparse==0.3.1
toml==0.10.0
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.2.4
websockets==8.1
whichcraft==0.6.1
wrapt==1.12.1
yarl==1.4.2
zope.event==4.4
zope.interface==5.1.0
